requirements.txt
langchain~=0.0.331
openai==0.27.8
tiktoken
requests
pydantic
pinecone-client
